feat: finish task

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        int division = toBeDivided / 5;
+        double division = toBeDivided / 5;
         System.out.println(division);
     }
 }
